+++ ccdproc/combiner.py
@@ -127,11 +127,12 @@ class Combiner:
     def weights(self, value):
         if value is not None:
             if isinstance(value, np.ndarray):
-                if value.shape == self.data_arr.data.shape:
-                    self._weights = value
-                else:
-                    raise ValueError(
-                        "dimensions of weights do not match data.")
+                if value.shape != self.data_arr.data.shape:
+                    if value.ndim != 1:
+                        raise ValueError("1D weights expected when shapes of the data and weights differ.")
+                    if value.shape[0] != self.data_arr.data.shape[0]:
+                        raise ValueError("Length of weights not compatible with specified axis.")
+                self._weights = value
             else:
                 raise TypeError("weights must be a numpy.ndarray.")
         else: