+++ ccdproc/tests/test_combiner.py
@@ -109,6 +109,17 @@ def test_weights_shape():
         c.weights = ccd_data.data
 
 
+def test_1Dweights():
+    ccd_list = [CCDData(np.zeros((10, 10)), unit=u.adu),
+                CCDData(np.zeros((10, 10)) - 1000, unit=u.adu),
+                CCDData(np.zeros((10, 10)) + 1000, unit=u.adu)]
+
+    c = Combiner(ccd_list)
+    c.weights = np.array([1, 5, 10])
+    ccd = c.average_combine()
+    np.testing.assert_almost_equal(ccd.data, 312.5)
+
+
 # test the min-max rejection
 def test_combiner_minmax():
     ccd_list = [CCDData(np.zeros((10, 10)), unit=u.adu),